requirements.txt
dataclasses
torch
deepspeed
transformers
peft
accelerate
pandas
pillow
fastparquet
requests
numpy
scipy